+++ src/main/java/ind/maiweiqi/hyperliquid_fill_quality/fill_quality_engine/FillQualityStats.java
@@ -10,7 +10,7 @@ public class FillQualityStats {
 
     // Lifetime counters
     private long totalTicks;
-    private long totalVolume;
+    private double totalVolume;
     private long totalWindows;
 
     // Rolling statistics
